// backend/src/common/enums/event-status.enum.ts
export enum EventStatus {
  STANDBY = 'STATUS_STANDBY',
  CHECKIN = 'STATUS_CHECKIN',
  ICEBREAKER = 'STATUS_ICEBREAKER',
  LOTTERY_READY = 'STATUS_LOTTERY_READY',
  LOTTERY_RUNNING = 'STATUS_LOTTERY_RUNNING',
  GAME_SHAKE = 'STATUS_GAME_SHAKE',
  GAME_MATCH = 'STATUS_GAME_MATCH',
  ENDED = 'STATUS_ENDED',
}